renamed files and adjusted home page layout.
