+++ frontend/src/components/event-invitation.jsx
