+++ frontend/src/pages/event-details.jsx
